news-summarizer/requirements.txt
feedparser
requests
streamlit
python-dateutil
nltk
duckduckgo-search
geocoder
pytz